# src/definitions/swagger.yaml
basePath: /
definitions:
    IUser:
        description: ""
        properties:
            id:
                type: string
                description: ""
            email:
                type: string
                description: ""
            firstname:
                type: string
                description: ""
            lastname:
                type: string
                description: ""
            username:
                type: string
                description: ""
            mobileNumber:
                type: string
                description: ""
            createdAt:
                type: string
                format: date-time
                description: ""
            updatedAt:
                type: string
                format: date-time
                description: ""
        type: object
        required:
            - id
            - email
            - firstname
            - lastname
            - username
            - mobileNumber
            - createdAt
            - updatedAt
...
    ISpecialBet:
        description: ""
        properties:
            id:
                type: string
                description: ""
            key:
                type: string
                description: ""
        type: object
        required:
            - id
            - key
...
    ISport:
        description: ""
        properties:
            id:
                type: string
                description: ""
            czName:
                type: string
                description: ""
            engName:
                type: string
                description: ""
            value:
                type: string
                description: ""
        type: object
        required:
            - id
            - czName
            - engName
            - value
    ITeam:
        description: ""
        properties:
            id:
                type: string
                description: ""
            czName:
                type: string
                description: ""
            engName:
                type: string
                description: ""
            value:
                type: string
                description: ""
            shortcut:
                type: string
                description: ""
            sport:
                type: string
                description: ""
            league:
                type: string
                description: ""
        type: object
        required:
            - id
            - czName
            - engName
            - value
            - shortcut
            - sport
            - league
    IUserBet:
        description: ""
        properties:
            id:
                type: string
                description: ""
            matchId:
                type: string
                description: ""
            userId:
                type: string
                description: ""
            dateTime:
                type: string
                format: date-time
                description: ""
            homeScore:
                type: string
                description: ""
            awayScore:
                type: string
                description: ""
            homeWinner:
                type: boolean
                description: ""
            scorer:
                type: string
                description: ""
        type: object
        required:
            - id
            - matchId
            - userId
            - dateTime
            - homeScore
            - awayScore
            - homeWinner
            - scorer
...
info:
    license:
        name: MIT
    title: tipapp-api
    version: 0.0.1
paths:
...
    '/api/users/{id}':
        get:
            operationId: UsersGetUser
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUser'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: UsersUpdateUser
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUser'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: user
                    required: true
                    schema:
                        type: object
        delete:
            operationId: UsersDeleteUser
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
...
    /api/specBets:
        get:
            operationId: SpecialBetsGetSpecialBets
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/ISpecialBet'
            description: ""
            parameters: []
        post:
            operationId: SpecialBetsCreateSpecialBet
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ISpecialBet'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: specialBet
                    required: true
                    schema:
                        type: object
    '/api/specBets/{id}':
        get:
            operationId: SpecialBetsGetSpecialBet
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ISpecialBet'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: SpecialBetsUpdateSpecialBet
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ISpecialBet'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: specialBet
                    required: true
                    schema:
                        type: object
        delete:
            operationId: SpecialBetsDeleteSpecialBet
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
...
    /api/sports:
        get:
            operationId: SportsGetSports
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/ISport'
            description: ""
            parameters: []
        post:
            operationId: SportsCreateSport
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ISport'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: sport
                    required: true
                    schema:
                        type: object
    '/api/sports/{id}':
        get:
            operationId: SportsGetSport
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ISport'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: SportsUpdateSport
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ISport'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: sport
                    required: true
                    schema:
                        type: object
        delete:
            operationId: SportsDeleteSport
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
    /api/teams:
        get:
            operationId: TeamsGetTeams
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/ITeam'
            description: ""
            parameters: []
        post:
            operationId: TeamsCreateTeam
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ITeam'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: team
                    required: true
                    schema:
                        type: object
    '/api/teams/{id}':
        get:
            operationId: TeamsGetTeam
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ITeam'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: TeamsUpdateTeam
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ITeam'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: team
                    required: true
                    schema:
                        type: object
        delete:
            operationId: TeamsDeleteTeam
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
    /api/userBets:
        get:
            operationId: UserBetsGetUserBets
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/IUserBet'
            description: ""
            parameters: []
        post:
            operationId: UserBetsCreateUserBet
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserBet'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: usetBet
                    required: true
                    schema:
                        type: object
    '/api/userBets/{id}':
        get:
            operationId: UserBetsGetUserBet
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserBet'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: UserBetsUpdateUserBet
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUserBet'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: usetBet
                    required: true
                    schema:
                        type: object
        delete:
            operationId: UserBetsDeleteUserBet
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
...
swagger: '2.0'
securityDefinitions: {}